+++ src/errors/application/tls.rs
@@ -6,6 +6,7 @@
 use thiserror::Error;
 
 #[derive(Debug, Error, Clone, PartialEq, Eq)]
+#[non_exhaustive]
 pub enum TlsError {
     #[error("TLS record too short")]
     TooShort,